Decrease range sizes
The primary purpose of these types seems to be to test and ensure constant folding; decreasing the lengths by a few orders of magnitude will still show huge performance changes if constant folding changes. This change will simply make them terminate in a reasonable time in case of a constant-fold failure.