More view performance tests (#54)
* More view performance tests

* View compatibility with CartesianIndex for .4 & .5

Define compatibility methods in a similar manner to how they're defined for getindex on each version.

* Decrease range sizes

The primary purpose of these types seems to be to test and ensure constant folding; decreasing the lengths by a few orders of magnitude will still show huge performance changes if constant folding changes. This change will simply make them terminate in a reasonable time in case of a constant-fold failure.