random: don't benchmark functions with interpolated types (#129)
This fixes part of #124. E.g. When wanting to benchmark
`rand(Int)` and `rand(UInt)`, the following loop won't
measure what is expected:
`for T=(Int, UInt); @benchmarkable rand($T); end`
In order to benchmark the equivalent of
`@benchmarkable rand(Int); @benchmarkable rand(UInt)`,
we "value-fy" the types with `Val`.