Remove deprecations (and coming deprecations) (#190)
When a deprecation appears in the hot loop of a tuned benchmark, BaseBenchmarks basically becomes non-terminating.  So this PR anticipates some of the coming changes in JuliaLang/julia#24368.  I have also taken the opportunity to fix the few other deprecations I ran into -- none of those were in hot loops.

Note that I have significantly changed some of these algorithms in order to better take advantage of dot-fusion.  I think that is in the spirit of those specific benchmarks, but if desired I can happily revert that back to the most-minimal change required to avoid the deprecation.