Fix readstring benchmark properly, reuse testbuf again.
`perf_read!` does a `seekstart` at the beginning, but `perf_read!` was not used for the `readstring` benchmark, only the `read` benchmark. So if the `read` benchmark was done before the `readstring` benchmark, `testbuf` would definitely be at eof, as witnessed by the too-good-to-be-true 80 ns benchmark time.

The implementation in the previous commit was also incorrect, as there can be multiple evaluations per sample.

Also reuse `testbuf` again in response to https://github.com/JuliaCI/BaseBenchmarks.jl/pull/223#issuecomment-410464560.