inference: fixup `many_method_matches` benchmark
It turns out that the previous call signature `many_method_matches(::Float64)`
was just wrong and so the benchmark hasn't been functional, especially
when benchmarking the performance of optimization passes (since the
preparatory abstract interpretation was not really successful).

Fixed it up, and also sets up an assertion that checks that the inference
runs on each benchmarks as expected, in order to avoid this sort of
problems.